backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
mysql-connector-python==8.3.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI/ML Dependencies
torch==2.1.2
torchvision==0.16.2
Pillow==10.2.0
numpy==1.26.3
opencv-python-headless==4.9.0.80

# HTTP Client
httpx==0.26.0
